+++ Connecto.DataObjects/EntityFramework/EntityProduct.cs
@@ -16,7 +16,7 @@ namespace Connecto.DataObjects.EntityFramework
         public int StockInHand { get; set; }
         public double ContainsQty { get; set; }
         public int Quantity { get; set; }
-        public int QuantityLower { get; set; }
+        public double QuantityLower { get; set; }
         public int Reorderlevel { get; set; }
         public virtual EntityProductType ProductType { get; set; }
         public virtual EntityVendor Vendor { get; set; }
