+++ Connecto.DataObjects/EntityFramework/EntitySalesDetail.cs
@@ -17,7 +17,7 @@ namespace Connecto.DataObjects.EntityFramework
         public double UnitPrice { get; set; }
         public double SellingPrice { get; set; }
         public int Quantity { get; set; }
-        public int QuantityLower { get; set; }
+        public double QuantityLower { get; set; }
         public DateTime DateSold { get; set; }
         public virtual EntityCustomer Customer { get; set; }
     }
